wip: t_node receive long instead int
Validation of INT_MIN/INT_MAX will be in stack_a, so we
need change our t_node type.

+++ mandatory/init/linked_lists.c
@@ -1,6 +1,6 @@
 #include <push_swap.h>
 
-t_node	*create_node(int number)
+t_node	*create_node(long number)
 {
 	t_node	*node;
