+++ mandatory/push_swap.h
@@ -21,7 +21,7 @@
 typedef struct s_node	t_node;
 struct s_node
 {
-	int				number;
+	long				number;
 	const int		index;
 	int				position;
 	int				target_position;
@@ -30,7 +30,7 @@ struct s_node
 	t_node			*next;
 };
 
-t_node		*create_node(int number);
+t_node		*create_node(long number);
 t_node		*get_last_node(t_node **head);
 void		link_node(t_node **head, t_node *node);
 void		free_list(t_node **head);
